Stricter constraint for $ref for schemas
In JSON Schema, /definitions is just a common place (a best practice)
to store reused sub-schema definitions. The property names used there
do not matter in /definitions. They only matter for $refs.
And sub-schemas can be stored anywhere else (for example in /whatever)
as there is no constraint.

Swagger 2.0 is inspired from JSON Schema, but it adds meaning to
/definitions: it contains "models" and the name of the model matters
for code/doc generators.
Also, the locations for storage of free elements (such as schemas) is
strictly limited in Swagger to /definitions or to vendor extensions
(properties starting with 'x-' in limited locations).
So we have an additional constraint for $refs pointing to a schema in
the same file: they must point to something in /definitions (or to avoid
blocking extension, to something below a vendor extension).
This commit adapts the JSON Schema for Swagger 2.0 to enforce this rule.